pistachio: fix the compilation issues.
Fix below compilation warning.

[..]

tools/bimgtool.c:76:17: warning: ‘crc_x25’ defined but not used [-Wunused-function]
 static uint16_t crc_x25(uint16_t crc, void *void_buf, size_t size)
                 ^
board/imgtec/pistachio_bub/pistachio.c: In function 'board_eth_init':
board/imgtec/pistachio_bub/pistachio.c:113:19: warning: assignment discards 'const' qualifier from pointer target type
  enet_dtb_macaddr = get_dtb_macaddr(0);
                   ^
drivers/mtd/spi-nand/spi-nand-base.c: In function 'spi_nand_register':
drivers/mtd/spi-nand/spi-nand-base.c:469:12: warning: assignment discards 'const' qualifier from pointer target type
  mtd->name = snand->name;
            ^
drivers/net/designware.c: In function 'designware_eth_probe':
drivers/net/designware.c:572:2: warning: format '%lx' expects argument of type 'long unsigned int', but argument 3 has type 'phys_addr_t' [-Wformat=]
  debug("%s, iobase=%lx, priv=%p\n", __func__, pdata->iobase, priv);
  ^
board/imgtec/pistachio_bub/pistachio.c: In function 'board_eth_init':
board/imgtec/pistachio_bub/pistachio.c:113:19: warning: assignment discards 'const' qualifier from pointer target type
  enet_dtb_macaddr = get_dtb_macaddr(0);
                   ^

[..]

Add ethernet boot command and remove usb part number from usb boot
most usb drives don't seem to have part-number.

+++ drivers/net/designware.c
@@ -610,7 +610,8 @@ static int designware_eth_probe(struct udevice *dev)
 	}
 #endif
 
-	debug("%s, iobase=%x, priv=%p\n", __func__, iobase, priv);
+	debug("%s, iobase=%x, priv=%p\n", __func__,
+				(unsigned long int)iobase, priv);
 	priv->mac_regs_p = (struct eth_mac_regs *)iobase;
 	priv->dma_regs_p = (struct eth_dma_regs *)(iobase + DW_DMA_BASE_OFFSET);
 	priv->interface = pdata->phy_interface;
